move util fns to is own file
