+++ src/modules/BudgetStatement/schema/budgetStatement.ts
@@ -1,4 +1,11 @@
 import { gql, AuthenticationError } from "apollo-server-core";
+import {
+  convertDate,
+  getAnalyticsActuals,
+  getAnalyticsForecast,
+  getAnalyticsOnChain,
+  getAnalyticsOffChain
+} from "./utils.js";
 
 export const typeDefs = [
   gql`
@@ -1130,90 +1137,3 @@ export const resolvers = {
     },
   },
 };
-
-function convertDate(dateString: string) {
-  const [year, month] = dateString.split('-');
-
-  const formattedDate = `${year}/${month}`;
-
-  return formattedDate;
-}
-const getAnalyticsActuals = async (queryEngine: any, monthAndDay: string, ownerType: string, teamId: number) => {
-
-  const filter = {
-    start: "2020/01",
-    end: "2100/01",
-    granularity: 'total',
-    metrics: ['Actuals'],
-    dimensions: [
-      { name: 'report', select: `atlas/${ownerType}/${teamId}/${monthAndDay}`, lod: 5 }
-    ],
-    currency: 'DAI'
-  }
-  const results = await queryEngine.query(filter);
-
-  const result = results.map((s: any) => ({
-    actuals: s.rows.find((r: any) => r.metric == 'Actuals')?.value,
-  }))
-  return result;
-}
-
-const getAnalyticsForecast = async (queryEngine: any, monthAndDay: string, ownerType: string, teamId: number) => {
-
-  const filter = {
-    start: "2020/01",
-    end: "2100/01",
-    granularity: 'total',
-    metrics: ['Forecast'],
-    dimensions: [
-      { name: 'report', select: `atlas/${ownerType}/${teamId}/${monthAndDay}`, lod: 5 }
-    ],
-    currency: 'DAI'
-  }
-  const results = await queryEngine.query(filter);
-
-  const result = results.map((s: any) => ({
-    forecast: s.rows.find((r: any) => r.metric == 'Forecast')?.value
-  }))
-  return result;
-}
-
-const getAnalyticsOnChain = async (queryEngine: any, monthAndDay: string, ownerType: string, teamId: number) => {
-
-  const filter = {
-    start: "2020/01",
-    end: "2100/01",
-    granularity: 'total',
-    metrics: ['PaymentsOnChain'],
-    dimensions: [
-      { name: 'report', select: `atlas/${ownerType}/${teamId}/${monthAndDay}`, lod: 5 }
-    ],
-    currency: 'DAI'
-  }
-  const results = await queryEngine.query(filter);
-
-  const result = results.map((s: any) => ({
-    paymentsOnChain: s.rows.find((r: any) => r.metric == 'PaymentsOnChain')?.value,
-  }))
-  return result;
-}
-
-const getAnalyticsOffChain = async (queryEngine: any, monthAndDay: string, ownerType: string, teamId: number) => {
-
-  const filter = {
-    start: "2020/01",
-    end: "2100/01",
-    granularity: 'total',
-    metrics: ['PaymentsOffChainIncluded'],
-    dimensions: [
-      { name: 'report', select: `atlas/${ownerType}/${teamId}/${monthAndDay}`, lod: 5 }
-    ],
-    currency: 'DAI'
-  }
-  const results = await queryEngine.query(filter);
-
-  const result = results.map((s: any) => ({
-    paymentsOffChain: s.rows.find((r: any) => r.metric == 'PaymentsOffChainIncluded')?.value,
-  }))
-  return result;
-}
